Remove "RE" from var names in `wifi` package
Most package-scope `var`s are regular expressions, so "RE" is redundant.
And since the meaning is not immediately obvious, it can be frustrating
to figure out if it's important to understand.  Finally, it should be
easy in most editors to view the type definition for a variable, making
it easy to figure out these are regexes.

Signed-off-by: Paul Zhang <paulz@ucla.edu>